More fixes to report package formatting
