+++ Website/UrlExtensions.cs
@@ -63,12 +63,12 @@ namespace NuGetGallery
 
         public static string Package(this UrlHelper url, string id)
         {
-            return url.Package(id, null);
+            return url.Package(id, null, scheme: null);
         }
 
-        public static string Package(this UrlHelper url, string id, string version)
+        public static string Package(this UrlHelper url, string id, string version, string scheme = null)
         {
-            string result = url.RouteUrl(RouteName.DisplayPackage, new { id, version });
+            string result = url.RouteUrl(RouteName.DisplayPackage, new { id, version }, protocol: scheme);
 
             // Ensure trailing slashes for versionless package URLs, as a fix for package filenames that look like known file extensions
             return version == null ? EnsureTrailingSlash(result) : result;
@@ -119,9 +119,9 @@ namespace NuGetGallery
             return url.Action(actionName: "UploadPackage", controllerName: MVC.Packages.Name);
         }
 
-        public static string User(this UrlHelper url, User user)
+        public static string User(this UrlHelper url, User user, string scheme = null)
         {
-            string result = url.Action(MVC.Users.Profiles(user.Username));
+            string result = url.Action(MVC.Users.Profiles(user.Username), protocol: scheme);
             return EnsureTrailingSlash(result);
         }
